% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAIN.R
\name{bptest}
\alias{bptest}
\title{Breusch-Pagan Test}
\usage{
bptest(formula, varformula = NULL, studentize = TRUE, data = list())
}
\arguments{
\item{formula}{a symbolic description for the model to be tested (or a fitted \code{"lm"} object).}

\item{varformula}{a formula describing only the potential explanatory variables
for the variance (no dependent variable needed). By default the same
explanatory variables are taken as in the main regression model.}

\item{studentize}{logical. If set to \code{TRUE} Koenker's studentized version of the test statistic will be used.}

\item{data}{an optional data frame containing the variables in the model. By default the variables are taken from the environment which \code{bptest} is called from.}
}
\value{
A list with class \code{"htest"} containing the following components:
 \item{statistic}{the value of the test statistic.}
 \item{p.value}{the p-value of the test.}
 \item{parameter}{degrees of freedom.}
 \item{method}{a character string indicating what type of test was performed.}
 \item{data.name}{a character string giving the name(s) of the data.}
}
\description{
Performs the Breusch-Pagan test against heteroskedasticity. This function is exported from the "lmtest" package (Achim Zeileis & Torsten Hothorn, 2002).
This function provides a way to test the heteroskedasticity identification condition (5) in Lemma 1 of Tchetgen Tchetgen et al. (2017).
}
\details{
The Breusch-Pagan test fits a linear regression model to the residuals
of a linear regression model (by default the same explanatory variables are taken as in the main regression
model) and rejects if too much of the variance is explained by the additional explanatory variables. Under H0 the test statistic of the Breusch-Pagan test follows a
chi-squared distribution with \code{parameter} (the number of regressors without the constant in the model) degrees of freedom.
}
\examples{
# the following packages are needed to simulate data
library("msm")
library("MASS")

nIV=10; N=500; beta=1;
phi=rep(-0.5,nIV); gamma=rep(-2,nIV); alpha=rep(-0.5,nIV);
lambda0=1; lambda1=rep(0.5,nIV);
Gn = mvrnorm(N,rep(0,nIV),diag(rep(1,nIV)))

G  = (Gn>0)*1;
U = as.vector(phi\%*\%t(G))+rnorm(N);
A = as.vector(gamma\%*\%t(G)) +U + rnorm(N,mean=0,sd=abs(lambda0+as.vector(lambda1\%*\%t(G))));
Y = as.vector(alpha\%*\%t(G)) + beta*A + U + rnorm(N);

bptest(A~G);
}
\references{
T.S. Breusch & A.R. Pagan (1979),
A Simple Test for Heteroscedasticity and Random Coefficient Variation.
\emph{Econometrica} \bold{47}, 1287--1294

R. Koenker (1981), A Note on Studentizing a Test for Heteroscedasticity.
\emph{Journal of Econometrics} \bold{17}, 107--112.

W. Kramer & H. Sonnberger (1986),
\emph{The Linear Regression Model under Test}. Heidelberg: Physica

Achim Zeileis & Torsten Hothorn (2002), Diagnostic Checking in Regression Relationships. \emph{R News} \bold{2(3)}, 7-10. \url{https://CRAN.R-project.org/doc/Rnews/}

Tchetgen Tchetgen, E., Sun, B. and Walter, S. (2017). \href{https://arxiv.org/abs/1709.07779}{The GENIUS Approach to Robust Mendelian Randomization Inference}. arXiv e-prints.
}
