% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAIN.R
\name{genius_mulA}
\alias{genius_mulA}
\title{MR GENIUS under multiplicative exposure model}
\usage{
genius_mulA(Y, A, G, alpha = 0.05, lower = -10, upper = 10)
}
\arguments{
\item{Y}{A numeric vector of outcomes.}

\item{A}{A numeric vector of exposures (binary values should be coded in 0/1).}

\item{G}{A numeric matrix of instruments; each column stores values for one instrument (a numeric vector if only a single instrument is available).}

\item{alpha}{Significance level for confidence interval (default value=0.05).}

\item{lower}{The lower end point of the causal effect interval to be searched (default value=-10).}

\item{upper}{The upper end point of the causal effect interval to be searched (default value=10).}
}
\value{
A "genius" object containing the following items:
\item{beta.est}{The point estimate of the causal effect (on the additive scale) of the exposure on the outcome.}
\item{beta.var}{The corresponding estimated variance.}
\item{ci}{The corresponding Wald-type confidence interval at specified significance level.}
\item{pval}{The p-value for two-sided Wald test of null causal effect (on the additive scale) of the exposure on the outcome.}
}
\description{
Implements MR GENIUS under a multiplicative exposure model.
}
\details{
This function implements the estimator given in Lemma 3 of Tchetgen Tchetgen et al (2017), under a multiplicative exposure model. By default, the log ratio term in equation (9) is modelled as a linear combination of the main effects of all available instruments.
}
\examples{
#the following package is needed to simulate data
library("MASS")
nIV=10; N=2000; beta=0.5; 
gamma=rep(0.5,nIV); alpha=rep(0.5,nIV);phi=rep(0.05,nIV);
Gn = mvrnorm(N,rep(0,nIV),diag(rep(1,nIV)))
G  = (Gn>0)*1;
U = as.vector(phi\%*\%t(G))+rnorm(N);
#exposure generated from negative binomial distribution
A = rnbinom(N,size=10,mu = exp(as.vector(gamma\%*\%t(G)) +0.1*U)) 
Y = as.vector(alpha\%*\%t(G)) + beta*A + U + rnorm(N);

genius_mulA(Y,A,G);


}
\references{
Tchetgen Tchetgen, E., Sun, B. and Walter, S. (2017). \href{https://arxiv.org/abs/1709.07779}{The GENIUS Approach to Robust Mendelian Randomization Inference}. arXiv e-prints.
}
